exclusion_angle

Return the inclusion ring defined for the radar location and a minimum and maximum distance

exclusion_angle(
  location = c(r_lon, r_lat),
  angles = b_corners[1, ],
  distmax = mad,
  crs = loc_epsg,
  verbose = F
)

Arguments

location

coordinate, vector of radar coordinates, or spatial point. Input coordinates in WGS84.

angles

vector of minimum and maximum angles in degrees to exclude

distmax

proj4string to transform to for calculations in a metric system

crs

crs to perform calculations in

Value

a polygon with inclusion donut

Details

With the argument exclusion_angle for which we expect structural interference. Here we have two types of interference:

The turbine the radar is situated on blocks one side of the radar window

The area overlapping with the vertical radar beam, which mostly has mixed tracks we did not include in this study. Therefore this area is not relevant here We remove the blocked corners by creating a triangle using the radar coordinates, the corner angles, and maximum distance.

Examples

angles <- c(287,30)
distmax <- 2500 # distance in meters
crs <- 23095 # metric CRS in which calculations are performed

# location as vector

location <- c(4.185345, 52.42783)
blocked_triangle <- exclusion_angle(location, angles, distmax, crs)
ggplot(blocked_triangle) + geom_sf() + geom_sf(data=st_sfc(st_point(location),crs=4326))
#> Error in ggplot(blocked_triangle): could not find function "ggplot"

angles <- list(c(287,30),c(115,135))
blocked_triangles <- lapply(angles,function(x) exclusion_angle(location,angles=x,distmax,crs))
ggplot() +
geom_sf(data=blocked_triangles[[1]]) +
geom_sf(data=blocked_triangles[[2]]) +
geom_sf(data=st_sfc(st_point(location),crs=4326))
#> Error in ggplot(): could not find function "ggplot"