exclusion_location

Return the inclusion ring defined for the radar location and a minimum and maximum distance

exclusion_geom(x = turbines, buffer = NULL, crs = loc_epsg, verbose = F)

Arguments

x

coordinate, vector of radar coordinates, or spatial point. Input coordinates in WGS84.

buffer

proj4string to transform to for calculations in a metric system

crs

crs to perform calculations in

y

vector of minimum and maximum angles in degrees to exclude

Value

a polygon with inclusion donut

Details

Examples

turbines
#> Simple feature collection with 471 features and 9 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 484442.6 ymin: 5713772 xmax: 702752.2 ymax: 5996107
#> Projected CRS: ETRS89 / UTM zone 31N
#> First 10 features:
#>    OBJECTID  UTM_X   UTM_Y          NAAM     STATUS OPMERKING GEBIEDNR
#> 1         1 578089 5808005 Luchterduinen In gebruik   Turbine        9
#> 2         2 580868 5807450 Luchterduinen In gebruik   Turbine        9
#> 3         3 580393 5807703 Luchterduinen In gebruik   Turbine        9
#> 4         4 579918 5807955 Luchterduinen In gebruik   Turbine        9
#> 5         5 579443 5808207 Luchterduinen In gebruik   Turbine        9
#> 6         6 578969 5808460 Luchterduinen In gebruik   Turbine        9
#> 7         7 578494 5808712 Luchterduinen In gebruik   Turbine        9
#> 8         8 581238 5808169 Luchterduinen In gebruik   Turbine        9
#> 9         9 580641 5808486 Luchterduinen In gebruik   Turbine        9
#> 10       10 580145 5808749 Luchterduinen In gebruik   Turbine        9
#>    TURBINE_NR                             EIGENAAR               geometry
#> 1          30 Luchterduinen: Q10 Offshore Wind BV. POINT (578089 5808005)
#> 2          31 Luchterduinen: Q10 Offshore Wind BV. POINT (580868 5807450)
#> 3          32 Luchterduinen: Q10 Offshore Wind BV. POINT (580393 5807703)
#> 4          33 Luchterduinen: Q10 Offshore Wind BV. POINT (579918 5807955)
#> 5          34 Luchterduinen: Q10 Offshore Wind BV. POINT (579443 5808207)
#> 6          35 Luchterduinen: Q10 Offshore Wind BV. POINT (578969 5808460)
#> 7          36 Luchterduinen: Q10 Offshore Wind BV. POINT (578494 5808712)
#> 8          37 Luchterduinen: Q10 Offshore Wind BV. POINT (581238 5808169)
#> 9          38 Luchterduinen: Q10 Offshore Wind BV. POINT (580641 5808486)
#> 10         39 Luchterduinen: Q10 Offshore Wind BV. POINT (580145 5808749)
x<- turbines
buffer <- 100 # distance in meter
crs <- 23095 # metric CRS in which calculations are performed

p <- exclusion_geom(x,buffer,crs,verbose=T)
#> Error in st_transform(input_x, crs = crs): object 'input_x' not found

x <- st_buffer(x$geometry,200)
#> Error in st_buffer(x$geometry, 200): could not find function "st_buffer"
p <- exclusion_geom(x,buffer,crs,verbose=T)
#> Error in st_transform(input_x, crs = crs): object 'input_x' not found