inclusion_ring

Return the inclusion ring defined for the radar location and a minimum and maximum distance

inclusion_ring(location, distance, crs, verbose = F)

Arguments

location

coordinate, vector of radar coordinates, or spatial point. Input coordinates in WGS84.

distance

vector of minimum and maximum distance in meters to include

crs

metric crs to perform calculations in

Value

a polygon with inclusion ring in coordinate reference system of the input

Details

Examples

if (FALSE) {
distance <- c(2500,1000) # distance in meters
crs <- 23095 # metric CRS in which calculations are performed

# location as vector

location <- c(4.185345, 52.42783)
ring <- inclusion_ring(location, distance,crs)
ggplot(ring) + geom_sf() + geom_sf(data=st_sfc(st_point(location),crs=4326))

# location as geometry

# EPSG:4326
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
ring <- inclusion_ring(location, distance, crs)
ggplot(ring) + geom_sf() + geom_sf(data=location)

# EPSG:23095
location <- st_sfc(st_point(c(444696.9, 5809144)),crs=crs)
ring <- inclusion_ring(location, distance, crs)
ggplot(ring) + geom_sf() + geom_sf(data=location)
}