roi

Return the area of interest. Three filters (inclusion_donut, exclusion_angle, exclusion_location) can be specified.

roi(
  location = NULL,
  distance = NULL,
  excl_angles = NULL,
  excl_geom = NULL,
  excl_buffer = NULL,
  crs = loc_epsg,
  verbose = T
)

Arguments

location

coordinate, vector of radar coordinates, or spatial point. Input coordinates in WGS84.

crs

crs to perform calculations in. If different from location coordinate, the location coordinate will reprojected

inclusion_ring

vector of minimum and maximum distance in meters to include

exclusion_angle

vector, list or data.frame of angle-pairs to exclude from analysis

exclusion_location

sf-object with locations which can reliably be excluded

exclusion_location_buffer

buffer in unit of crs?

Value

a polygon with inclusion zones

Details

With the argument exclusion_angle for which we expect structural interference. Here we have two types of interference:

The turbine the radar is situated on blocks one side of the radar window

The area overlapping with the vertical radar beam, which mostly has mixed tracks we did not include in this study. Therefore this area is not relevant here We remove the blocked corners by creating a triangle using the radar coordinates, the corner angles, and maximum distance.

Examples

# inclusion ring

# location as vector
location <- c(4.185345, 52.42783)
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

# location as geom
# 4326
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
#> Error in st_sfc(st_point(c(4.185345, 52.42783)), crs = 4326): could not find function "st_sfc"
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found
# 3035
location <- st_sfc(st_point(c(3925960, 3273397)),crs=3035)
#> Error in st_sfc(st_point(c(3925960, 3273397)), crs = 3035): could not find function "st_sfc"
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found
## distance

# one distance value: no exclusion of area near radar
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
#> Error in st_sfc(st_point(c(4.185345, 52.42783)), crs = 4326): could not find function "st_sfc"
area_of_inclusion <-  roi(location=location,
                         distance=2500,
                         crs=23095)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found
# two distance values: no exclusion of area near radar
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
#> Error in st_sfc(st_point(c(4.185345, 52.42783)), crs = 4326): could not find function "st_sfc"
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

## exclusion angles
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
#> Error in st_sfc(st_point(c(4.185345, 52.42783)), crs = 4326): could not find function "st_sfc"

# exclusion angles can be a list, vector or matrix

# list
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=list(c(287,30),c(115,135)))
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

# vector
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135))
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

# matrix
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=matrix(c(287,30,115,135),nrow=2,ncol=2,byrow=T))
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

# do not exclude areas near the radar
area_of_inclusion <-  roi(location=location,
                         distance=c(2500),
                         crs=23095,
                         excl_angles=matrix(c(287,30,115,135),nrow=2,ncol=2,byrow=T))
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

## exclusion areas

# without buffer points will not be excluded
location <- st_sfc(st_point(c(4.185345, 52.42783)),crs=4326)
#> Error in st_sfc(st_point(c(4.185345, 52.42783)), crs = 4326): could not find function "st_sfc"

area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=turbines)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

# buffer outside the function
polygons <- st_buffer(st_transform(turbines$geometry,23095),100)
#> Error in st_buffer(st_transform(turbines$geometry, 23095), 100): could not find function "st_buffer"
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=polygons)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found
# buffer within the function
area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=turbines,
                         excl_buffer=100
)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

# vary the buffer distance based on distance from radar
dist <- st_distance(location,turbines$geometry)
#> Error in st_distance(location, turbines$geometry): could not find function "st_distance"
weight <- as.vector(dist/max(dist) * 100)
#> Error in max(dist): invalid 'type' (closure) of argument

area_of_inclusion <-  roi(location=location,
                         distance=c(2500,1000),
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=turbines,
                         excl_buffer=weight
)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

# apply a buffer around multiple locations by using lists
area_of_inclusion <-  roi(location=location,
                         distance=2500,
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=list(turbines,location),
                         excl_buffer=list(100,1000)
)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found

# apply a buffer around multiple locations by using lists and varying buffer weight within a single dataset
dist <- st_distance(location,turbines$geometry)
#> Error in st_distance(location, turbines$geometry): could not find function "st_distance"
weight <- as.vector(dist/max(dist) * 100)
#> Error in max(dist): invalid 'type' (closure) of argument
area_of_inclusion <-  roi(location=location,
                         distance=2500,
                         crs=23095,
                         excl_angles=c(287,30,115,135),
                         excl_geom=list(turbines,location),
                         excl_buffer=list(weight,1000)
)
#> Error in paste0("coordinates are converted to sfc_POINT EPSG:4326.\nwarning: If input coordinates are not 4326, please input a geometry with the correct CRS as location-argument. \ne.g. st_sfc(st_point(location),crs=",     loc_epsg, ")"): object 'loc_epsg' not found